Change method for Google TTS
The package previously used for Google's text-to-speech
seemed to sometimes cause some problem where it froze
and where the beginning of the sound file would repeat
after the sound being played, even after restarting the
program. Restart of the system's audio driver was needed
to reset it. Now, another package is used to generate
sound files from Google's text-to-speech, and a separate
package is used to play them.