Show lemmas and their translations in lookup
For a word that is looked up, its lemmas (the basic word, e.g.
infinitive of a verb or nominative of a noun) are shown in the
translation section. If a personal translation has been added to a
lemma, it is shown as well. The lemmas associated to a word and their
translations can be editied by pressing "U" on the keyboard. This
enables an editable textfield where one lemma is listed at each line
(use shift + enter to make a new line). A translation can be associated
with a lemma by writing "<lemma>: <translation>". If there is no colon
at the line, the lemma is not associated with any translation.