Improve handling of line breaks
Keep adding an extra line break between paragraphs if all paragraphs are
spaced by a single line break. However, if the text has sufficiently
many double line breaks, so that we can assume that the paragraphs are
spaced by double line breaks, single line breaks are replaced by spaces.
The purpose of this is to handle texts where the paragraphs have line
breaks to keep the text width limited (which is the case for some
e-books).