Adapt font after operating system
When running the program on Linux, the font size for QTextEdit-based
widgets (which are all texts except for the titles in the side field)
became too large. This would presumably happen for Windows as well. Now,
the font size is automatically decreased when the operating system is
Linux or Windows.

Furthermore, if the font specified in the config file isn't available, a
default font (Helvetica) is used instead.