Add support for GPT machine translations
Enable the possibility to use one of OpenAI's GPT models for machine
translations (as a complement to fetching translations from Wiktionary)
instead of Google Translate. Google Translate is still default, but GPT
can be used instead by changing the config.json file. This option
requires an OpenAI API key.

The GPT translator first finds the dictionary form of the selected word,
and then returns one or several translations of this base form.