Add lemma from machine translation
When a word is looked up and no translation is found in Wiktionary, so
that a machine translation is added instead, the lemma (dictionary form)
obtained from this translation (if available) is added to the lemmas of
the looked-up word.

Furthermore, a lemma from the natural language processing model is added
only if a Wiktionary entry is found for this lemma, to prevent adding
false lemmas (which sometimes are produced by the NLP model).